+++ src/classes/extensions/ScreenExt.js
@@ -5,15 +5,27 @@ import DOMSize from "../add-ons/DOMSize.js"
 /**
  * Screen extension
  *
- * @property {DOMSize} size Screen size
- * @property {DOMSize} resolution Screen resolution
- *
  * @hideconstructor
  * @memberof extensions
  */
 class ScreenExt extends Extension {
 	static properties = {
+		/**
+		 * Screen size
+		 *
+		 * @name extensions.ScreenExt#size
+		 * @type {DOMSize}
+		 * @readonly
+		 */
 		size: {get: () => new DOMSize(Math.floor(screen.width), Math.floor(screen.height)), configurable: true},
+
+		/**
+		 * Screen resolution
+		 *
+		 * @name extensions.ScreenExt#resolution
+		 * @type {DOMSize}
+		 * @readonly
+		 */
 		resolution: {get: () => new DOMSize(Math.floor(screen.width * devicePixelRatio), Math.floor(screen.height * devicePixelRatio)), configurable: true}
 	}
 }